AlmostCompleteGame.java
Game class almost perfect (i.e. good enough) in terms of code. Only problem is that I want the frame to be unresizeable, but doing frame.resizeable(false) creates these small borders of gaps in the bottom, middle of board and scoreboard and on the right side, which I believe is due to factoring in the borders of the frame.

The other thing is that I've neglected the comments so far, thought correcting those should be easy.