Actual working Game (oops) until Ben's done
I know Ben is working on making changing Game at the moment but I just realised that the version that was on here was not the version I was working with on Wednesday, so I've changed it to that. I must have forgot to upload it to github on wednesday.